feat: GameManager restructure
details: Renamed showAllCards flag to debugShowOpponentInfo; opponent chances of winning are only displayed if this flag is on. GameStateDto is now used to communicate to the GUI the state of the game. StateType indicates what the GameManager needs. PlayerAction is now InputAction. Added Ping action to request the current state of the game (not fully working).