feat: GameManager now covers edgecases and game end. I should have used TDD here, would've helped a ton
details: GameManager correctly handles if less than 2 players can bet during a round, the game skips to showdown. If at the start of a round, only 1 player can bet they are considered the winner of the game. If  2 players are betting, player (1) with more money checks, and then the other player goes all in, player 1 is given the chance to respond.